requirements.txt
asgiref==3.6.0
Django==3.2.16
Pillow==9.4.0
python-dateutil==2.8.2
pytz==2022.7.1
reportlab==3.5.50
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
